+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Better handle mixed dtype sources in the multi source ([#1616](../../pull/1616))
 - Add a utility function to minimize caching ([#1617](../../pull/1617))
 - Allow passing through converter options when writing a zarr sink to a non-zarr format ([#1618](../../pull/1618))
+- Do less calculations when applying affine transforms in the multi-source ([#1619](../../pull/1619))
 
 ### Bug Fixes